Code Review
You did a fantastic job with this. Your code was great to read overall, if not repetitive.

But that can be forgiven when you've got a working app that's well executed.

I spent a lot of time reviewing your code to implement a way to DRY up the code and remove repetition. The code posted is working but I'm glad to answer any questions you have about it.

Highlights:
* Verbose code
* API integration
* (mostly) Responsive!
* Solid structure and design

To do:
* Make sure you spend more time with APIs that you use in your projects. There were a few neat shortcuts available to you and you'll see some examples of them in the review!
* Refactoring: This is a skill that take time to hone. It is the process of rewriting your code to DRY it up and improve it overall.